on corrige les betises

+++ donnees.h
@@ -5,6 +5,6 @@ double recup_L(float L);
 double recup_H(float H);
 int recup_M(float M);
 int recup_n(float n);
-int recup_nb_iter(float nb_iter);
+int recup_nb_iter(int nb_iter);
 
 #endif
